# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/flights/airports": {
      "get": {
        "operationId": "FlightsController_airports",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "کاتالوگ فرودگاه‌ها برای فرم افزودن پرواز",
        "tags": [
          "flights"
        ]
      },
      "post": {
        "operationId": "FlightsController_createAirport",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAirportDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "افزودن شهر/فرودگاه جدید به کاتالوگ",
        "tags": [
          "flights"
        ]
      }
    },
...
    "/flights": {
      "post": {
        "operationId": "FlightsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFlightDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "افزودن پرواز جدید (مودال طراحی)",
        "tags": [
          "flights"
        ]
      }
    },
    "/flights/schedules": {
      "post": {
        "operationId": "FlightsController_createSchedule",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScheduleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "ثبت برنامه تکرارشونده پرواز (RRULE) و ساخت پروازهای آینده",
        "tags": [
          "flights"
        ]
      },
      "get": {
        "operationId": "FlightsController_listSchedules",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "فهرست برنامه‌های تکرارشونده پرواز",
        "tags": [
          "flights"
        ]
      }
    },
...
    "/flights/{instanceId}": {
      "get": {
        "operationId": "FlightsController_detail",
        "parameters": [
          {
            "name": "instanceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "مودال جزئیات پرواز: تفکیک واقعی کانال فروش + مجموع درآمد",
        "tags": [
          "flights"
        ]
      }
    },
    "/flights/{instanceId}/plan": {
      "patch": {
        "operationId": "FlightsController_plan",
        "parameters": [
          {
            "name": "instanceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanFlightDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "نرخ‌گذاری و تخصیص پرواز آینده — نرخ قابل فروش همچنان با تأیید مدیر عامل (فاز ۶)",
        "tags": [
          "flights"
        ]
      }
    },
    "/flights/{instanceId}/aircraft": {
      "patch": {
        "operationId": "FlightsController_changeAircraft",
        "parameters": [
          {
            "name": "instanceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeAircraftTypeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "تغییر نوع هواپیمای پرواز — رد با ۴۰۹ اگر ظرفیت جدید کمتر از رزروهای قطعی/لاک‌شده باشد",
        "tags": [
          "flights"
        ]
      }
    },
    "/flights/{instanceId}/fare-rules": {
      "get": {
        "operationId": "FlightsController_listFareRules",
        "parameters": [
          {
            "name": "instanceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "فهرست کلاس‌های نرخی این پرواز",
        "tags": [
          "flights"
        ]
      },
      "post": {
        "operationId": "FlightsController_createFareRule",
        "parameters": [
          {
            "name": "instanceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFareRuleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "ایجاد کلاس نرخی — رد با ۴۰۰ اگر مجموع صندلی تخصیص‌یافته از ظرفیت کابین بیشتر شود",
        "tags": [
          "flights"
        ]
      }
    },
    "/flights/{instanceId}/fare-rules/{ruleId}": {
      "patch": {
        "operationId": "FlightsController_updateFareRule",
        "parameters": [
          {
            "name": "instanceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ruleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFareRuleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "ویرایش کلاس نرخی",
        "tags": [
          "flights"
        ]
      },
      "delete": {
        "operationId": "FlightsController_deleteFareRule",
        "parameters": [
          {
            "name": "instanceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ruleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "حذف کلاس نرخی — رد با ۴۰۹ اگر رزرو فعالی از آن استفاده کند",
        "tags": [
          "flights"
        ]
      }
    },
    "/flights/{instanceId}/allotments": {
      "get": {
        "operationId": "FlightsController_listAllotments",
        "parameters": [
          {
            "name": "instanceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "فهرست سهمیه‌های آژانس این پرواز",
        "tags": [
          "flights"
        ]
      },
      "post": {
        "operationId": "FlightsController_createAllotment",
        "parameters": [
          {
            "name": "instanceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAllotmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "تخصیص سهمیه به آژانس — رد با ۴۰۰ اگر مجموع سهمیه‌ها از سقف کلی آژانس‌های پرواز بیشتر شود",
        "tags": [
          "flights"
        ]
      }
    },
    "/flights/{instanceId}/allotments/{allotmentId}": {
      "delete": {
        "operationId": "FlightsController_deleteAllotment",
        "parameters": [
          {
            "name": "instanceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "allotmentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "حذف سهمیه آژانس — رد با ۴۰۹ اگر آژانس رزرو فعالی داشته باشد",
        "tags": [
          "flights"
        ]
      }
    },
...
  },
  "info": {
    "title": "blujet API",
    "description": "blujet airline platform — internal API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateAirportDto": {
        "type": "object",
        "properties": {
          "cityFa": {
            "type": "string",
            "description": "نام شهر",
            "example": "وان"
          },
          "code": {
            "type": "string",
            "description": "کد IATA فرودگاه",
            "example": "VAS"
          },
          "tz": {
            "type": "string",
            "description": "منطقه زمانی IANA",
            "example": "Asia/Tehran"
          }
        },
        "required": [
          "cityFa",
          "code"
        ]
      },
      "CreateFlightDto": {
        "type": "object",
        "properties": {
          "originCode": {
            "type": "string",
            "description": "کد فرودگاه مبدأ",
            "example": "THR"
          },
          "destCode": {
            "type": "string",
            "description": "کد فرودگاه مقصد",
            "example": "DXB"
          },
          "flightNo": {
            "type": "string",
            "description": "شماره پرواز",
            "example": "EP-901"
          },
          "departureAt": {
            "type": "string",
            "description": "زمان پرواز (UTC ISO — تبدیل جلالی در فرانت)"
          },
          "capacity": {
            "type": "number",
            "description": "ظرفیت صندلی",
            "example": 180
          },
          "basePriceIrr": {
            "type": "string",
            "description": "قیمت پایه (ریال)",
            "example": "38000000"
          }
        },
        "required": [
          "originCode",
          "destCode",
          "flightNo",
          "departureAt",
          "capacity",
          "basePriceIrr"
        ]
      },
      "CreateScheduleDto": {
        "type": "object",
        "properties": {
          "originCode": {
            "type": "string",
            "description": "کد فرودگاه مبدأ",
            "example": "THR"
          },
          "destCode": {
            "type": "string",
            "description": "کد فرودگاه مقصد",
            "example": "MHD"
          },
          "flightNo": {
            "type": "string",
            "description": "شماره پرواز",
            "example": "BJ-410"
          },
          "rrule": {
            "type": "string",
            "description": "الگوی تکرار RRULE (RFC 5545)",
            "example": "FREQ=WEEKLY;BYDAY=SA,MO,WE"
          },
          "depTime": {
            "type": "string",
            "description": "ساعت حرکت (UTC)",
            "example": "07:30"
          },
          "capacity": {
            "type": "number",
            "description": "ظرفیت هر پرواز",
            "example": 146
          },
          "daysAhead": {
            "type": "number",
            "description": "چند روز آینده از حالا ساخته شود (پیش‌فرض ۳۰)",
            "example": 30
          }
        },
        "required": [
          "originCode",
          "destCode",
          "flightNo",
          "rrule",
          "depTime",
          "capacity"
        ]
      },
      "PlanFlightDto": {
        "type": "object",
        "properties": {
          "priceIrr": {
            "type": "string",
            "description": "نرخ برنامه‌ریزی (ریال)",
            "example": "39000000"
          },
          "agencySeats": {
            "type": "number",
            "description": "تخصیص صندلی آژانس",
            "example": 60
          },
          "saleStartsAt": {
            "type": "string",
            "description": "شروع بازه فروش (UTC ISO) — خالی/حذف یعنی بدون محدودیت"
          },
          "saleEndsAt": {
            "type": "string",
            "description": "پایان بازه فروش (UTC ISO) — خالی/حذف یعنی بدون محدودیت"
          }
        },
        "required": [
          "priceIrr",
          "agencySeats"
        ]
      },
      "ChangeAircraftTypeDto": {
        "type": "object",
        "properties": {
          "aircraftType": {
            "type": "string",
            "description": "نوع هواپیمای جدید",
            "example": "Boeing 737"
          },
          "stepUpChallengeId": {
            "type": "string",
            "description": "از POST /auth/step-up/request (scope: PRICE_CAPACITY_CHANGE)"
          },
          "stepUpCode": {
            "type": "string",
            "example": "482913"
          }
        },
        "required": [
          "aircraftType",
          "stepUpChallengeId",
          "stepUpCode"
        ]
      },
      "CreateFareRuleDto": {
        "type": "object",
        "properties": {
          "cabin": {
            "type": "string",
            "enum": [
              "ECONOMY",
              "BUSINESS"
            ]
          },
          "classCode": {
            "type": "string",
            "description": "کد کلاس نرخی",
            "example": "Y"
          },
          "priceIrr": {
            "type": "string",
            "description": "قیمت (ریال)",
            "example": "30000000"
          },
          "seatsAllocated": {
            "type": "number",
            "description": "تعداد صندلی تخصیص‌یافته",
            "example": 20
          },
          "taxIrr": {
            "type": "string",
            "description": "مالیات/عوارض (ریال)",
            "example": "0"
          },
          "refundable": {
            "type": "boolean"
          },
          "changeable": {
            "type": "boolean"
          },
          "baggageAllowanceKg": {
            "type": "number",
            "example": 20
          },
          "validFrom": {
            "type": "string",
            "description": "شروع بازه اعتبار (UTC ISO)"
          },
          "validUntil": {
            "type": "string",
            "description": "پایان بازه اعتبار (UTC ISO)"
          },
          "allowedChannels": {
            "type": "array",
            "description": "کانال‌های مجاز — خالی یعنی همه کانال‌ها",
            "items": {
              "type": "string",
              "enum": [
                "SYSTEM",
                "CHARTER",
                "AGENCY"
              ]
            }
          }
        },
        "required": [
          "cabin",
          "classCode",
          "priceIrr",
          "seatsAllocated"
        ]
      },
      "UpdateFareRuleDto": {
        "type": "object",
        "properties": {
          "priceIrr": {
            "type": "string"
          },
          "seatsAllocated": {
            "type": "number"
          },
          "taxIrr": {
            "type": "string"
          },
          "refundable": {
            "type": "boolean"
          },
          "changeable": {
            "type": "boolean"
          },
          "baggageAllowanceKg": {
            "type": "number"
          },
          "validFrom": {
            "type": "string"
          },
          "validUntil": {
            "type": "string"
          },
          "allowedChannels": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "SYSTEM",
                "CHARTER",
                "AGENCY"
              ]
            }
          }
        }
      },
      "CreateAllotmentDto": {
        "type": "object",
        "properties": {
          "agencyId": {
            "type": "string",
            "description": "شناسه کاربری آژانس"
          },
          "seatsAllocated": {
            "type": "number",
            "description": "تعداد صندلی تخصیص‌یافته",
            "example": 10
          },
          "type": {
            "type": "string",
            "enum": [
              "SOFT",
              "HARD"
            ],
            "default": "HARD"
          },
          "releaseAt": {
            "type": "string",
            "description": "موعد آزادسازی خودکار (فقط برای نوع SOFT، UTC ISO)"
          },
          "contractPriceIrr": {
            "type": "string",
            "description": "نرخ قراردادی این آژانس (ریال)"
          }
        },
        "required": [
          "agencyId",
          "seatsAllocated"
        ]
      },
...
    }
  }
}